Refactor to use modules more and cleanup params RES/MOD key
