+++ ecs_composex/cloudmap/cloudmap_stack.py
@@ -10,94 +10,38 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from troposphere.servicediscovery import PrivateDnsNamespace
+
+from ..resources_import import import_record_properties
+from ..vpc.vpc_params import VPC_ID
+
 if TYPE_CHECKING:
     from ecs_composex.mods_manager import XResourceModule
     from ecs_composex.common.settings import ComposeXSettings
 
-import warnings
 from copy import deepcopy
 
-from compose_x_common.aws.cloudmap import get_all_dns_namespaces
 from compose_x_common.compose_x_common import keyisset, set_else_none
 from troposphere import GetAtt, Ref
-from troposphere.servicediscovery import PrivateDnsNamespace
 
-from ecs_composex.common import add_update_mapping, build_template, setup_logging
+from ecs_composex.cloudmap.cloudmap_helpers import (
+    detect_duplicas,
+    lookup_service_discovery_namespace,
+    resolve_lookup,
+)
+from ecs_composex.common import LOG, add_update_mapping, build_template
 from ecs_composex.common.stacks import ComposeXStack
-from ecs_composex.exceptions import ComposeBaseException, IncompatibleOptions
-from ecs_composex.resources_import import import_record_properties
-from ecs_composex.vpc.vpc_params import VPC_ID
-
-from ..compose.x_resources.environment_x_resources import AwsEnvironmentResource
-from ..compose.x_resources.helpers import (
+from ecs_composex.compose.x_resources.environment_x_resources import (
+    AwsEnvironmentResource,
+)
+from ecs_composex.compose.x_resources.helpers import (
     set_lookup_resources,
     set_new_resources,
     set_resources,
 )
-from .cloudmap_params import (
-    LAST_DOT_RE,
-    MAPPINGS_KEY,
-    MOD_KEY,
-    PRIVATE_DNS_ZONE_ID,
-    PRIVATE_DNS_ZONE_NAME,
-    PRIVATE_NAMESPACE_ID,
-    RES_KEY,
-    ZONES_PATTERN,
-)
+
+from .cloudmap_params import MOD_KEY, PRIVATE_DNS_ZONE_NAME, PRIVATE_NAMESPACE_ID
 from .cloudmap_x_resources import handle_resource_cloudmap_settings
-
-LOG = setup_logging()
-
-
-def lookup_service_discovery_namespace(zone, session, ns_id=None):
-    """
-    Function to find and get the PrivateDnsNamespace properties needed by other resources
-
-    :param PrivateNamespace zone:
-    :param boto3.session.Session session:
-    :param str ns_id:
-    :return: The properties we need
-    :rtype: dict
-    """
-    client = session.client("servicediscovery")
-    try:
-        namespaces = get_all_dns_namespaces(session)
-        if zone.zone_name not in [z["Name"] for z in namespaces]:
-            raise LookupError(
-                "No private namespace found for zone", zone.name, zone.zone_name
-            )
-        zone_r = None
-        if not ns_id:
-            for l_zone in namespaces:
-                if zone.zone_name == l_zone["Name"]:
-                    the_zone = l_zone
-                    zone_r = client.get_namespace(Id=the_zone["Id"])
-                    break
-        else:
-            zone_r = client.get_namespace(Id=ns_id)
-        if not zone_r:
-            raise LookupError(
-                f"{zone.module.res_key}.{zone.name} - Failed to lookup {zone.zone_name}"
-            )
-        properties = zone_r["Namespace"]["Properties"]
-        if zone_r["Namespace"]["Type"] == "HTTP":
-            raise TypeError(
-                "Unsupported CloudMap namespace HTTP. "
-                "Only DNS namespaces, private or public, are supported"
-            )
-        return {
-            PRIVATE_DNS_ZONE_ID: properties["DnsProperties"]["HostedZoneId"],
-            PRIVATE_DNS_ZONE_NAME: LAST_DOT_RE.sub(
-                "", properties["HttpProperties"]["HttpName"]
-            ),
-            PRIVATE_NAMESPACE_ID: zone_r["Namespace"]["Id"],
-        }
-    except client.exceptions.NamespaceNotFound:
-        LOG.error(f"Namespace not found for {zone.name}")
-        raise
-    except client.exceptions.InvalidInput:
-        LOG.error("Failed to retrieve the zone info")
-        raise
 
 
 class PrivateNamespace(AwsEnvironmentResource):
@@ -241,29 +185,39 @@ class PrivateNamespace(AwsEnvironmentResource):
             root_stack.stack_template.add_resource(self.stack)
 
 
-def resolve_lookup(lookup_resources, settings):
+class XStack(ComposeXStack):
     """
-    Lookup the ACM certificates in AWS and creates the CFN mappings for them
+    Root stack for x-cloudmap
 
-    :param list[HostedZone] lookup_resources: List of resources to lookup
     :param ecs_composex.common.settings.ComposeXSettings settings:
     """
-    if not keyisset(MAPPINGS_KEY, settings.mappings):
-        settings.mappings[MAPPINGS_KEY] = {}
-    for resource in lookup_resources:
-        resource.lookup_resource(
-            ZONES_PATTERN,
-            lookup_service_discovery_namespace,
-            PrivateDnsNamespace.resource_type,
-            "",
-        )
-        resource.init_outputs()
-        resource.generate_cfn_mappings_from_lookup_properties()
-        resource.generate_outputs()
-        settings.mappings[MAPPINGS_KEY].update(
-            {resource.logical_name: resource.mappings}
-        )
-    LOG.debug(settings.mappings[MAPPINGS_KEY])
+
+    _title = "AWS CloudMap Namespaces"
+
+    def __init__(
+        self, name: str, settings: ComposeXSettings, module: XResourceModule, **kwargs
+    ):
+        """
+        :param str name:
+        :param ecs_composex.common.settings.ComposeXSettings settings:
+        :param dict kwargs:
+        """
+        set_resources(settings, PrivateNamespace, module)
+        x_resources = settings.compose_content[module.res_key].values()
+        detect_duplicas(x_resources)
+        lookup_resources = set_lookup_resources(x_resources)
+        new_resources = set_new_resources(x_resources, supports_uses_default=True)
+        for resource in x_resources:
+            resource.stack = self
+        if new_resources:
+            stack_template = build_template(self._title)
+            super().__init__(module.mapping_key, stack_template, **kwargs)
+            define_new_namespace(new_resources, stack_template)
+        else:
+            self.is_void = True
+        if lookup_resources:
+            resolve_lookup(lookup_resources, settings, module)
+        self.module_name = module.mod_key
 
 
 def define_new_namespace(new_namespaces, stack_template):
@@ -314,78 +268,3 @@ def define_new_namespace(new_namespaces, stack_template):
         stack_template.add_resource(namespace.cfn_resource)
         namespace.init_outputs()
         namespace.generate_outputs()
-
-
-def x_cloud_lookup_and_new_vpc(settings, vpc_stack):
-    """
-    Function to ensure there is no x-cloudmap.Lookup resource and Compose-X is creating a new VPC.
-    The Namespace (CloudMap PrivateNamespace) cannot span across multiple VPC
-
-    :param ecs_composex.common.settings.ComposeXSettings settings:
-    :param vpc_stack: The VPC Stack
-    :raises: IncompatibleOptions
-    """
-    lookup_namespaces = [
-        namespace
-        for namespace in settings.x_resources
-        if isinstance(namespace, PrivateNamespace) and namespace.lookup_properties
-    ]
-    if lookup_namespaces and not vpc_stack.is_void:
-        raise IncompatibleOptions(
-            "You cannot have Compose-X Create a new VPC and use x-cloudmap.Lookup."
-            " Use x-vpc to re-use the VPC the PrivateNamespace is attached to",
-            lookup_namespaces,
-        )
-
-
-def detect_duplicas(x_resources: list):
-    """
-    Function to ensure there is no multiple resources with the same zone name
-
-    :param list[PrivateNamespace] x_resources:
-    """
-
-    class DuplicateZoneName(ComposeBaseException):
-        pass
-
-    names = [res.zone_name for res in x_resources]
-    if len(names) > len(set(names)):
-        raise DuplicateZoneName(
-            "There is a duplicate of zone names. All names",
-            names,
-            "unique names",
-            set(names),
-        )
-
-
-class XStack(ComposeXStack):
-    """
-    Root stack for x-cloudmap
-
-    :param ecs_composex.common.settings.ComposeXSettings settings:
-    """
-
-    _title = "AWS CloudMap Namespaces"
-
-    def __init__(self, name: str, settings, module, **kwargs):
-        """
-        :param str name:
-        :param ecs_composex.common.settings.ComposeXSettings settings:
-        :param dict kwargs:
-        """
-        set_resources(settings, PrivateNamespace, module)
-        x_resources = settings.compose_content[module.res_key].values()
-        detect_duplicas(x_resources)
-        lookup_resources = set_lookup_resources(x_resources)
-        new_resources = set_new_resources(x_resources, supports_uses_default=True)
-        for resource in x_resources:
-            resource.stack = self
-        if new_resources:
-            stack_template = build_template(self._title)
-            super().__init__(name, stack_template, **kwargs)
-            define_new_namespace(new_resources, stack_template)
-        else:
-            self.is_void = True
-        if lookup_resources:
-            resolve_lookup(lookup_resources, settings)
-        self.module_name = module.mod_key
