+++ ecs_composex/efs/efs_params.py
@@ -5,14 +5,12 @@
 import re
 from os import path
 
-from ecs_composex.common import NONALPHANUM
 from ecs_composex.common.cfn_params import Parameter
 from ecs_composex.common.ecs_composex import X_KEY
 from ecs_composex.vpc.vpc_params import SG_ID_TYPE
 
 MOD_KEY = path.basename(path.dirname(path.abspath(__file__)))
 RES_KEY = f"{X_KEY}{MOD_KEY}"
-MAPPINGS_KEY = NONALPHANUM.sub("", MOD_KEY)
 
 
 FS_REGEXP = re.compile(
