--- ecs_composex/events/events_params.py
@@ -1,10 +1,0 @@
-#  -*- coding: utf-8 -*-
-# SPDX-License-Identifier: MPL-2.0
-# Copyright 2020-2022 John Mille <john@compose-x.io>
-
-from os import path
-
-from ecs_composex.common.ecs_composex import X_KEY
-
-MOD_KEY = path.basename(path.dirname(path.abspath(__file__)))
-RES_KEY = f"{X_KEY}{MOD_KEY}"
