+++ pyproject.toml
@@ -65,6 +65,7 @@ tbump = "6.5.0"
 sphinx-jsonschema = "^1.17.2"
 boto3-stubs = { extras = ["acm", "ecs", "cloudformation", "essential"], version = "^1.21.18" }
 pyclean = "^2.1.0"
+sphinx-autodoc-typehints = "^1.17.0"
 
 [tool.poetry.scripts]
 ecs-compose-x = "ecs_composex.cli:main"
